+++ src/vixMntOperation.cpp
@@ -2,8 +2,8 @@
 
 VixMntOpRead::VixMntOpRead(
         const char* fileName_,
-        size_t bufsize_,
-        off_t offsize_
+        uint64 bufsize_,
+        uint64 offsize_
         )
 {
     memcpy(fileName,fileName_,strlen(fileName));
@@ -13,8 +13,8 @@ VixMntOpRead::VixMntOpRead(
 
 VixMntOpWrite::VixMntOpWrite(
         const char* fileName_,
-        size_t bufsize_,
-        off_t offsize_
+        uint64 bufsize_,
+        uint64 offsize_
         )
 {
     memcpy(fileName,fileName_,strlen(fileName));
